+++ packages/react-dom/src/events/plugins/SelectEventPlugin.js
@@ -16,14 +16,10 @@ import {canUseDOM} from 'shared/ExecutionEnvironment';
 import {SyntheticEvent} from '../../events/SyntheticEvent';
 import isTextInputElement from '../isTextInputElement';
 import shallowEqual from 'shared/shallowEqual';
-import {enableEagerRootListeners} from 'shared/ReactFeatureFlags';
 
 import {registerTwoPhaseEvent} from '../EventRegistry';
 import getActiveElement from '../../client/getActiveElement';
-import {
-  getNodeFromInstance,
-  getEventListenerSet,
-} from '../../client/ReactDOMComponentTree';
+import {getNodeFromInstance} from '../../client/ReactDOMComponentTree';
 import {hasSelectionCapabilities} from '../../client/ReactInputSelection';
 import {DOCUMENT_NODE} from '../../shared/HTMLNodeType';
 import {accumulateTwoPhaseListeners} from '../DOMPluginEventSystem';
@@ -154,23 +150,6 @@ function extractEvents(
   eventSystemFlags: EventSystemFlags,
   targetContainer: EventTarget,
 ) {
-  if (!enableEagerRootListeners) {
-    const eventListenerSet = getEventListenerSet(targetContainer);
-    // Track whether all listeners exists for this plugin. If none exist, we do
-    // not extract events. See #3639.
-    if (
-      // If we are handling selectionchange, then we don't need to
-      // check for the other dependencies, as selectionchange is only
-      // event attached from the onChange plugin and we don't expose an
-      // onSelectionChange event from React.
-      domEventName !== 'selectionchange' &&
-      !eventListenerSet.has('onSelect') &&
-      !eventListenerSet.has('onSelectCapture')
-    ) {
-      return;
-    }
-  }
-
   const targetNode = targetInst ? getNodeFromInstance(targetInst) : window;
 
   switch (domEventName) {