+++ packages/shared/forks/ReactFeatureFlags.testing.www.js
@@ -48,7 +48,6 @@ export const enableNewReconciler = false;
 export const deferRenderPhaseUpdateToNextBatch = true;
 export const decoupleUpdatePriorityFromScheduler = false;
 export const enableDiscreteEventFlushingChange = true;
-export const enableEagerRootListeners = true;
 
 export const enableDoubleInvokingEffects = false;